Extract energy-devices-detail-graph-card series generation into a pure module
Move the per-collection-update series/legend generation out of the card
component into a pure `generateEnergyDevicesDetailGraphData` module. All
environment inputs (hass, computed styles, current time, and the untracked
dataset id suffix) are injected so the transform is deterministic and
benchmarkable. No behavior change.
